requirements.txt
# Основные библиотеки для работы с данными
numpy>=1.23.5,<2.0.0
opencv-python>=4.8.0.0
Pillow

# Базы данных
mysql-connector-python
pymilvus==2.3.4
python-dotenv

# Нейросетевой стек
deepface
tensorflow>=2.16.1
tf-keras

# Движок для ускорения на CPU (отлично подходит для Mac)
onnxruntime==1.16.3